refactor: format files with prettier, sort files and all components using hooks
