--- src/components/Footer.jsx
@@ -1,9 +1,0 @@
-import React from "react";
-
-const Footer = () => (
-    <div>
-        This is the footer!
-    </div>
-);
-
-export default Footer;
