+++ src/components/Layout/Footer.jsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+const Footer = () => <div>This is the footer!</div>;
+
+export default Footer;
