+++ src/components/Layout/PageLayout.jsx
@@ -1,7 +1,6 @@
 import React from "react";
-
-import Header from "../Header";
-import Footer from "../Footer";
+import Header from "./Header";
+import Footer from "./Footer";
 
 const PageLayout = ({ children }) => (
     <div className="layout">
