+++ src/components/Profile/Student.js
@@ -1,5 +1,5 @@
 export default class Student {
-    constructor({id, name, email, profileImage}) {
+    constructor({ id, name, email, profileImage }) {
         this.id = id;
         this.name = name;
         this.email = email;
